inchallah c'est final

+++ src/com/mco/GestionStock.java
@@ -3,6 +3,6 @@ package com.mco;
 import java.util.Map;
 
 public interface GestionStock {
-    public void achat(Article article, int quantite);
-    public void vendre(Article article, int quantite);
+    public void achat(Article article, double quantite);
+    public void vendre(Article article, double quantite);
 }
